+++ frontend/src/pageComponents/main/components/banner/Banner.styled.tsx
