+++ frontend/src/pageComponents/main/components/banner/index.tsx
