+++ frontend/src/pageComponents/main/components/instantwin/Instantwin.styled.tsx
