+++ frontend/src/pageComponents/main/components/instantwin/index.tsx
