+++ frontend/src/pageComponents/main/components/speedyattack/Speedyattack.styled.tsx
