+++ frontend/src/pageComponents/main/components/speedyattack/index.tsx
