+++ frontend/src/pageComponents/main/components/timeattack/Timeattack.styled.tsx
